requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.2
telethon>=1.34.0
celery>=5.3
redis>=5.0
sqlalchemy>=2.0
asyncpg>=0.29
alembic>=1.12
greenlet>=3.0
python-dotenv>=1.0
tenacity>=9.0
httpx>=0.27
cryptography>=42.0.5
aio-pika>=9.4

